railway_deploy/CHATMASIVO/requirements.txt
Flask==2.3.3
twilio==8.10.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pandas==2.1.4
openpyxl==3.1.2
